+++ src/cljs/cxengage_javascript_sdk/internal_utils.cljs
@@ -26,7 +26,7 @@
        (transform-keys camel/->kebab-case)))
 
 (defn build-api-url-with-params [url params]
-  (let [{:keys [tenant-id resource-id session-id entity-id entity-sub-id contact-id layout-id interaction-id artifact-id note-id artifact-file-id]} params]
+  (let [{:keys [tenant-id resource-id session-id entity-id entity-sub-id contact-id layout-id interaction-id artifact-id note-id artifact-file-id action-id]} params]
     (cond-> url
       tenant-id (clojure.string/replace #"tenant-id" (str tenant-id))
       resource-id (clojure.string/replace #"resource-id" (str resource-id))
@@ -36,6 +36,7 @@
       contact-id (clojure.string/replace #"contact-id" contact-id)
       layout-id (clojure.string/replace #"layout-id" layout-id)
       interaction-id (clojure.string/replace #"interaction-id" interaction-id)
+      action-id (clojure.string/replace #"action-id" action-id)
       artifact-id (clojure.string/replace #"artifact-id" artifact-id)
       note-id (clojure.string/replace #"note-id" note-id)
       artifact-file-id (clojure.string/replace #"artifact-file-id" artifact-file-id))))
@@ -95,6 +96,43 @@
                             {:headers {"Authorization" (str "Token " token)}})))]
      (ajax/ajax-request request)
      response-channel)))
+
+(defn file-api-request [request-map]
+  (let [response-channel (a/promise-chan)
+        {:keys [method url body]} request-map
+        request (merge {:uri url
+                        :method method
+                        :timeout 30000
+                        :handler #(let [normalized-response (normalize-response-stucture % false true)]
+                                    (a/put! response-channel normalized-response))
+                        :format (ajax/json-request-format)
+                        :response-format (ajax/json-response-format {:keywords? true})
+                        :body body}
+                       (when-let [token (state/get-token)]
+                         {:headers {"Authorization" (str "Token " token)}}))]
+    (ajax/ajax-request request)
+    response-channel))
+
+(defn get-artifact [interaction-id tenant-id artifact-id]
+  (let [url (str (state/get-base-api-url)
+                 (build-api-url-with-params
+                  "tenants/tenant-id/interactions/interaction-id/artifacts/artifact-id"
+                  {:tenant-id tenant-id
+                   :interaction-id interaction-id
+                   :artifact-id artifact-id}))
+        artifact-request {:method :get
+                          :url url}]
+    (api-request artifact-request)))
+
+(defn get-interaction-files [interaction-id]
+  (let [tenant-id (state/get-active-tenant-id)
+        url (-> "tenants/tenant-id/interactions/interaction-id/artifacts"
+                (build-api-url-with-params {:interaction-id interaction-id
+                                            :tenant-id tenant-id})
+                (#(str (state/get-base-api-url) %)))
+        file-request {:method :get
+                      :url url}]
+    (api-request file-request)))
 
 (defn format-response [response]
   (if (= :cljs (state/get-consumer-type))