+++ src/cljs/cxengage_javascript_sdk/state.cljs
@@ -161,9 +161,30 @@
         new-scripts (conj existing-scripts script)]
     (swap! sdk-state assoc-in [:interactions interaction-location interaction-id :scripts] new-scripts)))
 
+(defn get-script [interaction-id script-id]
+  (let [interaction-location (find-interaction-location interaction-id)
+        scripts (or (get-in @sdk-state [:interactions interaction-location interaction-id :scripts]) [])
+        filtered-script (first (filterv #(= script-id (:id (js->clj (js/JSON.parse (:script %)) :keywordize-keys true))) scripts))]
+    filtered-script))
+
 (defn add-email-artifact-data [interaction-id artifact-data]
   (let [interaction-location (find-interaction-location interaction-id)]
     (swap! sdk-state assoc-in [:interactions interaction-location interaction-id :email-artifact] artifact-data)))
+
+(defn add-email-reply-details-to-interaction [email-reply-details]
+  (let [{:keys [interaction-id]} email-reply-details
+        interaction-location (find-interaction-location interaction-id)]
+   (swap! sdk-state assoc-in [:interactions interaction-location interaction-id :email-reply-details] email-reply-details)))
+
+(defn get-all-email-attachments [interaction-id]
+  (let [interaction-location (find-interaction-location interaction-id)]
+    (get-in @sdk-state assoc-in [:interactions interaction-location interaction-id :email-reply-details :attachments])))
+
+(defn add-attachment-details-to-interaction [interaction-id attachment-details]
+  (let [interaction-location (find-interaction-location interaction-id)
+        all-current-attachments (get-all-email-attachments interaction-id)
+        new-attachments (conj all-current-attachments attachment-details)]
+    (swap! sdk-state assoc-in [:interactions interaction-location interaction-id :email-reply-details :attachments] new-attachments)))
 
 ;;;;;;;;;;;
 ;; Auth